Fix moving messages from threads to threads

+++ twake/backend/node/src/services/messages/services/engine/processors/channel-view/index.ts
@@ -46,46 +46,55 @@ export class ChannelViewProcessor {
       if (!message.resource.ephemeral) {
         //Publish message in corresponding channel
         if (message.created) {
+
           const pkPrefix = {
             company_id: participant.company_id,
             workspace_id: participant.workspace_id,
             channel_id: participant.id,
           };
 
-          await this.repository.save(
-            getInstance({
-              ...pkPrefix,
-              thread_id: thread.id,
-              message_id: message.resource.id,
-            }),
-          );
-
-          const reversed = await this.repositoryReversed.findOne({
+          //If a pointer exists it means the message already exists (it was probably moved and so we need to keep everything in place)
+          const existingPointer = await this.repository.findOne({
             ...pkPrefix,
-            thread_id: thread.id,
+            message_id: message.resource.id,
           });
+          if(!existingPointer){
 
-          if (reversed) {
-            const existingThreadRef = await this.repository.findOne({
-              ...pkPrefix,
-              message_id: reversed.message_id,
-            });
-            if (
-              existingThreadRef &&
-              `${existingThreadRef.thread_id}` === `${message.resource.thread_id}`
-            ) {
-              reversed.message_id = message.resource.id;
-              await this.repositoryReversed.save(reversed);
-              await this.repository.remove(existingThreadRef);
-            }
-          } else {
-            await this.repositoryReversed.save(
-              getInstanceReversed({
+            await this.repository.save(
+              getInstance({
                 ...pkPrefix,
                 thread_id: thread.id,
                 message_id: message.resource.id,
               }),
             );
+
+            const reversed = await this.repositoryReversed.findOne({
+              ...pkPrefix,
+              thread_id: thread.id,
+            });
+
+            if (reversed) {
+              const existingThreadRef = await this.repository.findOne({
+                ...pkPrefix,
+                message_id: reversed.message_id,
+              });
+              if (
+                existingThreadRef &&
+                `${existingThreadRef.thread_id}` === `${message.resource.thread_id}`
+              ) {
+                reversed.message_id = message.resource.id;
+                await this.repositoryReversed.save(reversed);
+                await this.repository.remove(existingThreadRef);
+              }
+            } else {
+              await this.repositoryReversed.save(
+                getInstanceReversed({
+                  ...pkPrefix,
+                  thread_id: thread.id,
+                  message_id: message.resource.id,
+                }),
+              );
+            }
           }
         }
       }
